+++ R/yocto-spectral2mspct.R
@@ -8,7 +8,7 @@
 #'
 #' @export
 #'
-yocto_spectral2mspct <- function(x, channels = "all") {
+yocto_spectral2mspct <- function(x) {
 
   if (is.null(channels)) {
     channels <- "all"
@@ -19,19 +19,14 @@ yocto_spectral2mspct <- function(x, channels = "all") {
   } else if (channels == "narrow") {
     channels <- paste("F", 1:8, sep = "")
   }
-  tidyr::pivot_longer(x,
-                      cols = !contains("time"),
-                      names_to = "channel",
-                      values_to = "counts") |>
-    dplyr::mutate(
-      channel = gsub("\\.avg$", "", channel),
-      w.length = AS7343_metadata()[channel]
-    ) |>
-    dplyr::arrange(time, w.length) -> z
-
-  if (! "all" %in% channels) {
-    z <- dplyr::filter(z, channel %in% channels)
-  }
+  z <- tidyr::pivot_longer(x,
+                           cols = grep("time", colnames(x), invert = TRUE),
+                           names_to = "channel",
+                           values_to = "counts")
+  z[["channel"]] <- gsub("\\.avg$", "", z[["channel"]])
+  z[["w.length"]] <- AS7343_metadata()[z[["channel"]]]
+  z <- z[order(z[["w.length"]])]
+  z[["time"]] <- factor(as.character(z[["time"]])) # better performance?
 
   photobiology::subset2mspct(z, member.class = "raw_spct", idx.var = "time")
 }